requirement.txt
vocode==0.1.110
python-dotenv==1.0.0